Exception handling - delete

+++ src/main/java/br/edu/toycenter/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import br.edu.toycenter.entities.User;
 import br.edu.toycenter.repositories.UserRepository;
+import br.edu.toycenter.services.exceptions.DatabaseException;
 import br.edu.toycenter.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,12 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
